refactor(api-client): move package schema and API verb interfaces for identity service
